--- a/packages/react-charts/src/components/Chart/Chart.tsx
+++ b/packages/react-charts/src/components/Chart/Chart.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import hoistNonReactStatics from 'hoist-non-react-statics';
 
 /* eslint-disable camelcase */
-import chart_legend_Margin from '@patternfly/react-tokens/dist/js/chart_legend_Margin';
+import chart_legend_Margin from '@patternfly/react-tokens/dist/esm/chart_legend_Margin';
 
 import {
   AnimatePropTypeInterface,