--- a/packages/react-charts/src/components/ChartTheme/styles/donut-utilization-styles.ts
+++ b/packages/react-charts/src/components/ChartTheme/styles/donut-utilization-styles.ts
@@ -1,6 +1,6 @@
 /* eslint-disable camelcase */
-import chart_donut_threshold_warning_Color from '@patternfly/react-tokens/dist/js/chart_donut_threshold_warning_Color';
-import chart_donut_threshold_danger_Color from '@patternfly/react-tokens/dist/js/chart_donut_threshold_danger_Color';
+import chart_donut_threshold_warning_Color from '@patternfly/react-tokens/dist/esm/chart_donut_threshold_warning_Color';
+import chart_donut_threshold_danger_Color from '@patternfly/react-tokens/dist/esm/chart_donut_threshold_danger_Color';
 
 // Donut utilization styles
 export const DonutUtilizationStyles = {