--- a/packages/react-charts/src/components/ChartTheme/styles/scatter-styles.ts
+++ b/packages/react-charts/src/components/ChartTheme/styles/scatter-styles.ts
@@ -1,6 +1,6 @@
 /* eslint-disable camelcase */
-import chart_scatter_active_size from '@patternfly/react-tokens/dist/js/chart_scatter_active_size';
-import chart_scatter_size from '@patternfly/react-tokens/dist/js/chart_scatter_size';
+import chart_scatter_active_size from '@patternfly/react-tokens/dist/esm/chart_scatter_active_size';
+import chart_scatter_size from '@patternfly/react-tokens/dist/esm/chart_scatter_size';
 
 // Donut styles
 export const ScatterStyles = {