--- a/packages/react-charts/src/components/ChartTheme/themes/axis-theme.ts
+++ b/packages/react-charts/src/components/ChartTheme/themes/axis-theme.ts
@@ -1,6 +1,6 @@
 /* eslint-disable camelcase */
-import chart_axis_grid_stroke_Color from '@patternfly/react-tokens/dist/js/chart_axis_grid_stroke_Color';
-import chart_axis_tick_stroke_Color from '@patternfly/react-tokens/dist/js/chart_axis_tick_stroke_Color';
+import chart_axis_grid_stroke_Color from '@patternfly/react-tokens/dist/esm/chart_axis_grid_stroke_Color';
+import chart_axis_tick_stroke_Color from '@patternfly/react-tokens/dist/esm/chart_axis_tick_stroke_Color';
 
 // Axis theme
 export const AxisTheme = {