--- a/packages/react-charts/src/components/ChartTheme/themes/dark/orange-color-theme.ts
+++ b/packages/react-charts/src/components/ChartTheme/themes/dark/orange-color-theme.ts
@@ -1,9 +1,9 @@
 /* eslint-disable camelcase */
-import chart_color_orange_100 from '@patternfly/react-tokens/dist/js/chart_color_orange_100';
-import chart_color_orange_200 from '@patternfly/react-tokens/dist/js/chart_color_orange_200';
-import chart_color_orange_300 from '@patternfly/react-tokens/dist/js/chart_color_orange_300';
-import chart_color_orange_400 from '@patternfly/react-tokens/dist/js/chart_color_orange_400';
-import chart_color_orange_500 from '@patternfly/react-tokens/dist/js/chart_color_orange_500';
+import chart_color_orange_100 from '@patternfly/react-tokens/dist/esm/chart_color_orange_100';
+import chart_color_orange_200 from '@patternfly/react-tokens/dist/esm/chart_color_orange_200';
+import chart_color_orange_300 from '@patternfly/react-tokens/dist/esm/chart_color_orange_300';
+import chart_color_orange_400 from '@patternfly/react-tokens/dist/esm/chart_color_orange_400';
+import chart_color_orange_500 from '@patternfly/react-tokens/dist/esm/chart_color_orange_500';
 import { ColorTheme } from '../color-theme';
 
 // Color scale