--- a/packages/react-charts/src/components/ChartTheme/themes/donut-utilization-theme.ts
+++ b/packages/react-charts/src/components/ChartTheme/themes/donut-utilization-theme.ts
@@ -1,12 +1,12 @@
 /* eslint-disable camelcase */
-import chart_donut_threshold_first_Color from '@patternfly/react-tokens/dist/js/chart_donut_threshold_first_Color';
-import chart_donut_threshold_second_Color from '@patternfly/react-tokens/dist/js/chart_donut_threshold_second_Color';
-import chart_donut_threshold_third_Color from '@patternfly/react-tokens/dist/js/chart_donut_threshold_third_Color';
-import chart_donut_utilization_dynamic_pie_Height from '@patternfly/react-tokens/dist/js/chart_donut_utilization_dynamic_pie_Height';
-import chart_donut_utilization_dynamic_pie_angle_Padding from '@patternfly/react-tokens/dist/js/chart_donut_utilization_dynamic_pie_angle_Padding';
-import chart_donut_utilization_dynamic_pie_Padding from '@patternfly/react-tokens/dist/js/chart_donut_utilization_dynamic_pie_Padding';
-import chart_donut_utilization_dynamic_pie_Width from '@patternfly/react-tokens/dist/js/chart_donut_utilization_dynamic_pie_Width';
-import chart_donut_utilization_static_pie_Padding from '@patternfly/react-tokens/dist/js/chart_donut_utilization_static_pie_Padding';
+import chart_donut_threshold_first_Color from '@patternfly/react-tokens/dist/esm/chart_donut_threshold_first_Color';
+import chart_donut_threshold_second_Color from '@patternfly/react-tokens/dist/esm/chart_donut_threshold_second_Color';
+import chart_donut_threshold_third_Color from '@patternfly/react-tokens/dist/esm/chart_donut_threshold_third_Color';
+import chart_donut_utilization_dynamic_pie_Height from '@patternfly/react-tokens/dist/esm/chart_donut_utilization_dynamic_pie_Height';
+import chart_donut_utilization_dynamic_pie_angle_Padding from '@patternfly/react-tokens/dist/esm/chart_donut_utilization_dynamic_pie_angle_Padding';
+import chart_donut_utilization_dynamic_pie_Padding from '@patternfly/react-tokens/dist/esm/chart_donut_utilization_dynamic_pie_Padding';
+import chart_donut_utilization_dynamic_pie_Width from '@patternfly/react-tokens/dist/esm/chart_donut_utilization_dynamic_pie_Width';
+import chart_donut_utilization_static_pie_Padding from '@patternfly/react-tokens/dist/esm/chart_donut_utilization_static_pie_Padding';
 
 // Donut utilization dynamic theme
 export const DonutUtilizationDynamicTheme = {