--- a/packages/react-charts/src/components/ChartTheme/themes/light/blue-color-theme.ts
+++ b/packages/react-charts/src/components/ChartTheme/themes/light/blue-color-theme.ts
@@ -1,9 +1,9 @@
 /* eslint-disable camelcase */
-import chart_color_blue_100 from '@patternfly/react-tokens/dist/js/chart_color_blue_100';
-import chart_color_blue_200 from '@patternfly/react-tokens/dist/js/chart_color_blue_200';
-import chart_color_blue_300 from '@patternfly/react-tokens/dist/js/chart_color_blue_300';
-import chart_color_blue_400 from '@patternfly/react-tokens/dist/js/chart_color_blue_400';
-import chart_color_blue_500 from '@patternfly/react-tokens/dist/js/chart_color_blue_500';
+import chart_color_blue_100 from '@patternfly/react-tokens/dist/esm/chart_color_blue_100';
+import chart_color_blue_200 from '@patternfly/react-tokens/dist/esm/chart_color_blue_200';
+import chart_color_blue_300 from '@patternfly/react-tokens/dist/esm/chart_color_blue_300';
+import chart_color_blue_400 from '@patternfly/react-tokens/dist/esm/chart_color_blue_400';
+import chart_color_blue_500 from '@patternfly/react-tokens/dist/esm/chart_color_blue_500';
 import { ColorTheme } from '../color-theme';
 
 // Color scale