--- a/packages/react-charts/src/components/ChartTheme/themes/threshold-theme.ts
+++ b/packages/react-charts/src/components/ChartTheme/themes/threshold-theme.ts
@@ -1,6 +1,6 @@
 /* eslint-disable camelcase */
-import chart_threshold_stroke_dash_array from '@patternfly/react-tokens/dist/js/chart_threshold_stroke_dash_array';
-import chart_threshold_stroke_Width from '@patternfly/react-tokens/dist/js/chart_threshold_stroke_Width';
+import chart_threshold_stroke_dash_array from '@patternfly/react-tokens/dist/esm/chart_threshold_stroke_dash_array';
+import chart_threshold_stroke_Width from '@patternfly/react-tokens/dist/esm/chart_threshold_stroke_Width';
 
 // Threshold theme
 export const ThresholdTheme = {