<!-- What changes are being made? Please link the issue being addressed. -->
**What**: Towards #6118. Would be good to get in before adding type checking to theme-patternfly-org downstream.

## Changes
- Upgrade to typescript@4.3.5 and tslib@2.3.1
  - Fix 2 type errors in ComposableTable b8679cb4a426a5ea94f5ecb6e2baac7a5567151a (no functional change)
  -  I've diffed the new and old dist folders and see only non-functional differences
    - Transformer uses single quotes instead of double quotes
    - Exports have been slightly tweaked to preserve variable names but are functionally the same. This is why a few snapshots are a little nicer in this PR
- Export types to `dist/esm` in addition to `dist/js`. This should make importing directly from `dist/esm` easier for consumers and also fixed a bug with incremental compiling
  - Also point `types` entry to `dist/esm/index.d.ts` and update rules to import from `/dist/esm` rather than `dist/js` (it was `dist/js` solely because the types were there...)
- Replace ttypescript (which doesn't work with TypeScript 4) with `ts-patch` (which does) to transform `dist/esm` imports to `dist/js` for our consumers
  - Have to check `module` instead of `target` now

### Performance
- Initial build times stay same at ~1m
- Incremental build times without changes increased from 5s to 15s
- Watch incremental build times with changes (more common) decreased from 12s to 9s.

<!-- Are there any upstream issues or separate issues you need to reference? -->
**Additional issues**: I'm willing to help consumers with problems this release should any arise...